% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_scc.R
\name{add_scc}
\alias{add_scc}
\title{Add state change counts}
\usage{
add_scc(tree_tbl, target)
}
\description{
This function counts the number of ancestral state changes between a
common ancestor and each tip which descends from this common ancestor. The 
function currently counts the number of state changes from root.
}
\examples{
\dontrun{
add_scc(tree_tbl, "k_serotype")
}
}
