% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_neighbors.R
\name{all_neighbors}
\alias{all_neighbors}
\title{List all neighbors of a country}
\usage{
all_neighbors(x, borders = NULL)
}
\arguments{
\item{x}{character; the two character ISO country code of a country}

\item{borders}{a data.frame which contains the country borders.}
}
\description{
This function uses an external database to list all neighbors of a country.
}
\note{
The database which is used by default does not include e.g. borders of
countries that are not universally recognised. However, you can customise
your database by adding or removing borders between countries. To do this,
use the \code{edit_borders} function and then use the resulting data frame
in the \code{borders} argument of this function.
}
\examples{
# USA has borders with Canada and Mexico
all_neighbors("US")

# Kosovo is not recovnised in the default database. As a results, Serbia has
# border with Albania
all_neighbors("RS")

# However, if we edit the borders database to recognise Kosoveo, Serbia will
# no longer have a border with Albania, insted it will have a border with
# Kosovo.
data(custom_country_borders)
updated_borders <- edit_borders(custom_country_borders)
all_neighbors("RS", borders = updated_borders)
}
\references{
https://github.com/wmgeolab/rgeoboundaries
}
