% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_middle.R
\name{date_middle}
\alias{date_middle}
\title{Convert date to numeric format for phylogenetic dating}
\usage{
date_middle(dates, out_format = "date")
}
\arguments{
\item{dates}{character; a full or partial date in "YYYY-MM_DD" like format.}

\item{out_format}{character; output format, either \code{"decimal"} or
\code{"date"}.}
}
\description{
This is a utility function that converts a date to a numeric value. This
format is commonly used in phylogenetic dating. If date is incomplete the
function returns the middle of the interval.
}
\examples{
date_middle("2021-18-11")
date_middle("1988-03")
date_middle("2000")
}
