% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_runif.R
\name{date_runif}
\alias{date_runif}
\title{Simulate random dates}
\usage{
date_runif(dates, out_format = "decimal")
}
\arguments{
\item{dates}{character; a vector of dates. Can be exact, e.g. 2000-01-01, or
incomplete, e.g. 2000-01 or 2000.}

\item{out_format}{character; output format, either \code{"decimal"} or
\code{"date"}.}
}
\value{
a vector of dates either in decimal format. If the input is complete,
the output will be the same as the input. If the input is incomplete, the
output will be a random date between the lower and upper boundaries of the
input.
}
\description{
This function returns exact dates and simulates random dates for incomplete
dates. For incomplete dates, the function will attempt to guess the lower and
upper boundaries from the input string.
}
\examples{
date_runif("2000-01-01")
date_runif("2000-01")
date_runif("2000")
}
