% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_upper.R
\name{date_upper}
\alias{date_upper}
\title{Convert date to numeric format for phylogenetic dating}
\usage{
date_upper(dates, out_format = "decimal")
}
\arguments{
\item{dates}{character; a full or partial date in "YYYY-MM_DD" like format.}

\item{out_format}{character; output format, either \code{"decimal"} or
\code{"date"}.}
}
\description{
This is a utility function that converts a date to a numeric value. This
format is commonly used in phylogenetic dating. If date is incomplete the
function returns the upper end of the interval.
}
\examples{
date_upper("2021-08-11")
date_upper("1988-03")
date_upper("2000")
date_upper("1900/1948")
date_upper("2003-03/2005-02")
}
