% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodist_matrix.R
\name{geodist_matrix}
\alias{geodist_matrix}
\title{Geographic distance matrix}
\usage{
geodist_matrix(
  df,
  id_var,
  lat_var = "lat",
  lon_var = "lon",
  focus_on = NULL,
  focus_by = NULL
)
}
\arguments{
\item{df}{data.frame; a data frame containing sample metadata}

\item{id_var}{character; a variable of the df we want to use as row and
column names. Must be unique for each row of the data frame.}

\item{lat_var}{character; a variable of the df we want to assess which
stores latitude coordinate.}

\item{lon_var}{character; a variable of the df we want to assess which
stores longitude coordinate.}

\item{focus_on}{character; one or more values of focus_by to focus on}

\item{focus_by}{character; a variable of the df to focus on}
}
\value{
a matrix
}
\description{
This function calculates a matrix which tells the geographic distance in km
between pairs of samples. This matrix is one of the "distance" matrices used
in risk calculations.
}
\examples{
df <- data.frame(
  assembly = LETTERS[1:5],
  continent = c("europe", "europe", "asia", "europe", "asia"),
  country = c("hungary", "germany", "china", "hungary", "laos"),
  lat = c(47.16249, 51.16569, 35.86166, 47.16249, 19.85627),
  lon = c(19.50330, 10.451526, 104.19540, 19.50330, 102.4955)
)
# no focus
geodist_matrix(df, id_var = "assembly")
# focus on Asian samples
geodist_matrix(
  df,
  id_var = "assembly",
  focus_by = "continent",
  focus_on = "asia"
)
}
