% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrca.R
\name{mrca}
\alias{mrca}
\title{Calculate times to most recent common ancestors}
\usage{
mrca(tree_tbl, method = "average", ...)
}
\arguments{
\item{tree_tbl}{tibble; the phylogenetic tree in tibble format.}

\item{method}{character; the method to be used for calculating time to the
most recent common ancestors. See details for more information.}

\item{...}{additional arguments used with \code{colldist()}.}
}
\value{
a tibble with five columns.
}
\description{
This function calculates phylogenetic (patristic) distances for all pairs of
tips within a phylogenetic tree.
}
\details{
The function implements three methods for calculating times to the
most recent common ancestors. \code{"first"} calculates times between the
ancestor and the earlier tip, \code("last") calculates times between the
ancestor and the more recent tip, \code("average") calculates the average of
the two which is identical to the half of the patristic distance.
}
\examples{
\dontrun {
# TODO put a tree_tbl into data so it can be imported.
mrca(tree_tbl)
}
}
