% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbors.R
\name{neighbors}
\alias{neighbors}
\title{Are two countries neighbors?}
\usage{
neighbors(x, y, custom_borders = NULL)
}
\arguments{
\item{x}{character; a vector of two characters ISO country codes}

\item{y}{character; a vector of two characters ISO country codes}

\item{custom_borders}{an object of class \code{custom_borders} which
modifies the default database.}
}
\description{
This function uses an external database to decide whether two countries have
land borders.
}
\note{
The database which is used by default does not include e.g. borders of
countries that are not universally recognised. However, you can customise
your database by adding or removing borders between countries. The
\code{custom_borders} argument lets you specify such database.
}
\examples{
# Hungary and Slovakia have a land border 
neighbors("HU", "SK")

# Hungary and Greece do not have a land border
neighbors("HU", "GR")
}
\references{
https://github.com/wmgeolab/rgeoboundaries
}
