% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_duplicates.R
\name{parse_duplicates}
\alias{parse_duplicates}
\title{Parse names for identical sequences}
\usage{
parse_duplicates(duplicates)
}
\arguments{
\item{duplicates}{character; a character vector.}
}
\value{
a list of duplicates. Each list element is a vector of ti
}
\description{
This is a utility function for parsing the names of identical sequences from
a text file which was populated by seqkit.This function can be used for
removing identical tips from a phylogenetic tree before operations that are
sensitive to zero branch lengths and adding them back after the operation.
}
\examples{
\dontrun{
# create an empty text file for storing duplicate sequence names
file.create(duplicates.txt)
# populate text file
seqkit rmdup -s seqs.fasta -D duplicates.txt -o seqs.nodup.fasta
# import duplicates
duplicates <- parse_duplicates("duplicates.txt")
}
}
