% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylodist.R
\name{phylodist}
\alias{phylodist}
\title{Phylogenetic distances}
\usage{
phylodist(tree_tbl)
}
\arguments{
\item{tree_tbl}{tibble; the phylogenetic tree in tibble format.}
}
\value{
a tibble with three columns.
}
\description{
This function calculates phylogenetic (patristic) distances for all pairs of
tips within a phylogenetic tree.
}
\examples{
\dontrun {
# TODO put a tree_tbl into data so it can be imported.
phylodist(tree_tbl)
}
}
