% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_snet.R
\name{plot_snet}
\alias{plot_snet}
\title{Plot ancestral state change network}
\usage{
plot_snet(snet, min_freq = 0.5)
}
\arguments{
\item{snet}{matrix; and adjacency matrix which contains frequencies
of ancestral state changes.}

\item{min_freq}{numeric; the lowest frequency of observed state changes 
which should be included on the plot.}

\item{cache_file}{character; the name of the cache file without the file
extension. If \code{NULL}, results are not cached.}
}
\description{
This function prepares a plot from an adjacency matrix which represents
directed ancestral state network along a phylogenetic tree.
}
\examples{
\dontrun{
plot_snet(snet, min_freq = 3)
}
}
