% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_statediff_mrca.R
\name{plot_statediff_mrca}
\alias{plot_statediff_mrca}
\title{Plot character state changes over phylogenetic distance}
\usage{
plot_statediff_mrca(tree_tbl, state_var, tips = NULL, binwidth = 2, ...)
}
\arguments{
\item{tree_tbl}{tibble; the phylogenetic tree in tibble format.}

\item{state_var}{character; the name of the variable which stores the
character state.}

\item{...}{additional arguments used with \code{mrca()}.}
}
\value{
a ggplot.
}
\description{
This convenience function integrates a number of other functions in the
package. It calculates most recent common ancestors and compares character
states between tips of a phylogenetic tree, and then plots the relative
frequency of character state changes over phylogenetic distance.
}
