% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree.R
\name{plot_tree}
\alias{plot_tree}
\title{Plot a phlyogenetic tree}
\usage{
plot_tree(
  tree_tbl,
  state_var,
  date_var = "collection_day",
  node_lab = NULL,
  export = TRUE
)
}
\arguments{
\item{tree_tbl}{tibble; the phylogenetic tree in tibble format.}

\item{state_var}{character; the name of the variable which stores the
character state.}

\item{date_var}{character; the name of the variable which stores sample
collection dates.}

\item{node_lab}{charcater; the name of the variable which stores node labels.
If \code{NULL}, node labels will be character states according to
\code{state_var}.}

\item{export}{logical; should the plot be exported?}
}
\description{
This function is used to plot a phylogenetic tree together with predicted
ancestral state changes that may have occurred along the branches of the
tree.
}
