% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scc.R
\name{scc}
\alias{scc}
\title{Count state changes between two states}
\usage{
scc(from, to, target, tree_tbl)
}
\arguments{
\item{from}{character; origin}

\item{to}{character; destination}

\item{target}{character; the target ancestral state to plot. Must match a
column name in \code{tree_tbl}.}

\item{tree_tbl}{tibble; dated tree with predicted ancestral states. This
tibble is automatically returned by the nextlow pipeline.}
}
\description{
Count state changes between two states
}
\examples{
\dontrun {
scc(from = "KL3", to = "KL2", target = "k_serotype", tree_tbl = tree_tbl)
}
}
