% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scp.R
\name{scp}
\alias{scp}
\title{Probability of state change}
\usage{
scp(from, to, from_which = NULL, to_which = NULL, split = "\\\\|")
}
\arguments{
\item{from}{character; origin}

\item{to}{character; destination}

\item{from_which}{character}

\item{to_which}{character}

\item{split}{character; the string separating multiple states within
\code{from} and/or \code{to}.}
}
\description{
This function calculates the combinatoric probability of state change
between two character strings.
}
\examples{
scp("KL2|KL3|KL49", "KL2|KL3")
scp("KL2|KL3|KL49", "KL2|KL3", from_which = "KL2", to_which = "KL3")
scp("KL2", "KL2|KL3")
}
