% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrink_matrix.R
\name{shrink_matrix}
\alias{shrink_matrix}
\title{Shrink matrix}
\usage{
shrink_matrix(matrix, small_matrix)
}
\arguments{
\item{matrix}{matrix; a matrix to shrink}

\item{small_matrix}{matrix; a matrix to shrink to}
}
\value{
a matrix
}
\description{
This function shrinks a matrix to have the same rows and columns as a smaller
matrix and also in the same order which is important for later calculations.
}
