% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snet.R
\name{snet}
\alias{snet}
\title{Prepare ancestral state network}
\usage{
snet(tree_tbl, target, cache_file = NULL)
}
\arguments{
\item{tree_tbl}{tibble; dated tree with predicted ancestral states. This
tibble is automatically returned by the nextlow pipeline.}

\item{target}{character; the target ancestral state to plot. Must match a
column name in \code{tree_tbl}.}

\item{cache_file}{character; the name of the cache file without the file
extension. If \code{NULL}, results are not cached.}
}
\value{
an adjacency matrix
}
\description{
This function prepares an adjacency matrix which represents a directed
ancestral state network along a phylogenetic tree. Nodes in this network
represent unique ancestral states, edges represent state changes between
ancestors and their descendants.
}
\examples{
\dontrun{
snet(tree_tbl, "k_serotype", cache_file = "adjmat_k_serotype")
}
}
\seealso{
plot_snet
}
