% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statediff.R
\name{statediff}
\alias{statediff}
\title{Compare states between tips}
\usage{
statediff(tree_tbl, state_var)
}
\arguments{
\item{tree_tbl}{tibble; the phylogenetic tree in tibble format.}

\item{state_var}{character; the name of the variable which stores the
character state.}
}
\value{
a tibble with three columns. The third column is 0 if the character
states are the same and 1 if they are different.
}
\description{
This function compares character states for all pairs of tips within a
phylogenetic tree.
}
\examples{
\dontrun {
# TODO put a tree_tbl into data so it can be imported.
statediff(tree_tbl, state_var = "k_serotype")
}
}
