% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempdist_matrix.R
\name{tempdist_matrix}
\alias{tempdist_matrix}
\title{Temporal distance matrix}
\usage{
tempdist_matrix(
  df,
  id_var,
  date_var,
  focus_on = NULL,
  focus_by = NULL,
  estimate_dates = "asis"
)
}
\arguments{
\item{df}{data.frame; a data frame containing sample metadata}

\item{id_var}{character; a variable of the df we want to use as row and
column names. Must be unique for each row of the data frame.}

\item{date_var}{character; a variable of the df we want to assess which
stores dates.}

\item{focus_on}{character; one or more values of focus_by to focus on}

\item{focus_by}{character; a variable of the df to focus on}

\item{estimate_dates}{character; strategy for estimating uncertain dates.
Either \code{"asis"} (no estimation), \code{"upper"}, \code{"middle"}, 
\code{"lower"} or \code{"runif"}.}
}
\value{
a matrix
}
\description{
This function calculates a matrix which tells the time difference between
sampling dates for pairs of samples. This matrix is one of the "distance"
matrices used in risk calculations.
}
\examples{
df <- data.frame(
  assembly = LETTERS[1:5],
  continent = c("europe", "europe", "asia", "europe", "asia"),
  country = c("hungary", "germany", "china", "hungary", "laos"),
  collection_date = c(
    "2000-01-01", "2001-01-01", "2005-01-01", "2000-06-15", "2005-06-15" 
  )
)
# no focus
tempdist_matrix(df, id_var = "assembly", date_var = "collection_date")
# focus on European samples
tempdist_matrix(
  df,
  id_var = "assembly",
  date_var = "collection_date",
  focus_by = "continent",
  focus_on = "asia"
)
}
\seealso{
[date_upper()],[date_middle()],[date_lower()], [date_runif()]
}
