% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_duplicates.R
\name{tidy_duplicates}
\alias{tidy_duplicates}
\title{Tidy duplicates for assembly subsets}
\usage{
tidy_duplicates(df, subset, id_var, duplicates)
}
\arguments{
\item{df}{data.frame; data set containing all samples}

\item{subset}{data.frame; a subset of the data set}

\item{id_var}{character; a variable in df and subset which contains the
duplicate names.}

\item{duplicates}{list; a list of duplicates. See \code{parse_duplicates()}
for more information.}
}
\value{
a list with two elements, \code{"subset"} and \code{"duplist"}.
\code{"subset"} contains the tidied subset data frame where any duplicates
are replaced by their reference assembly and \code{"duplist"} contains the
assemblies that have been removed in the process.
}
\description{
When subsampling an assembly data set that may contain assemblies with
identical sequences, it is important to eliminate any duplicates and ensure
that the duplicates are represented by their reference assembly, i.e. the
assembly that is present in the sequence data set.
}
